requirements.txt
django==4.0.4
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-db-comments==0.4.1
psycopg2-binary==2.9.3
coverage==6.4.1
drf-yasg==1.20.0
djangorestframework-simplejwt==5.2.0
elasticsearch==7.17.4
elasticsearch-dsl==7.4.0
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
gunicorn==20.1.0
gevent==21.12.0
django-filter==22.1